login

+++ app/Http/Controllers/TasksController.php
@@ -8,16 +8,32 @@ use App\Task;
 
 class TasksController extends Controller
 {
-  
+   public function __construct(){
+        $this->middleware('auth');
+    }
     public function index()
     {
-        $tasks =Task::all();
-
-        return view('tasks.index', [
-            'tasks' => $tasks,
-        ]);
+        
+        // ログインしていたら
+        $data = [];
+        if (\Auth::check()) { 
+            $user = \Auth::user();
+            // ログインユーザーのタスク一覧を取得して
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            
+            $data = [
+                    'user' => $user,
+                    'tasks' => $tasks,
+                ];
+            // tasks.indexを返す
+            return view('tasks.index', $data);
+      
+        // ログインしていなかったら
+        } else {
+        // サインアップとログインへのリンクのみのwelcomeを返す
+            return view('welcome');
+        }
     }
-
     /**
      * Show the form for creating a new resource.
      *
@@ -41,13 +57,22 @@ class TasksController extends Controller
 
     public function store(Request $request)
     {
+        
         $request->validate([
             'status' => 'required|max:10',  
+            'content' => 'required|max:255',
         ]);
         
+        // $request->user()->tasks()->create([
+        //     'status' => $request->status,
+        //     'content' => $request->content,
+        // ]);
+        
         $task = new Task;
-        $task->status = $request->status; 
+        $task->user_id = \Auth::id();
+        $task->status  = $request->status; 
         $task->content = $request->content;
+        
         $task->save();
         
         return redirect('/');
@@ -59,13 +84,16 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-      public function show($id)
+    public function show($id)
     {
         $task = Task::findOrFail($id);
 
-        return view('tasks.show', [
-            'task' => $task,
-        ]);
+        // もし、自分のタスクじゃなかったら、トップにリダイレクトする。
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
+        
+        return view('tasks.show',['task'=>$task,]);
     }
 
     /**
@@ -74,13 +102,15 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-     public function edit($id)
+    public function edit($id)
     {
         $task = Task::findOrFail($id);
 
-        return view('tasks.edit', [
-            'task' => $task,
-        ]);
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
+        
+        return  view('tasks.edit',['task'=>$task,]);
     }
 
     /**
@@ -90,12 +120,16 @@ class TasksController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-     public function update(Request $request, $id)
+    public function update(Request $request, $id)
     {
         $request->validate([
             'status' => 'required|max:10',  
+            'content' => 'required|max:255',  
         ]);
         $task = Task::findOrFail($id);
+        if (\Auth::id() !== $task->user_id) {
+            return redirect('/');
+        }
         $task->status = $request->status;
         $task->content = $request->content;
         $task->save();
@@ -112,8 +146,9 @@ class TasksController extends Controller
    public function destroy($id)
     {
         $task = Task::findOrFail($id);
+       if (\Auth::id() === $task->user_id) {
         $task->delete();
-
-        return redirect('/');
+        }
+        return back();
     }
 }
